added http status validation

+++ src/main/java/com/example/mbjpa/dto/AccountResponse.java
@@ -2,10 +2,12 @@ package com.example.mbjpa.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record AccountResponse(
         String actNo,
         String actType,
-        Double balance
+        BigDecimal balance
 ) {
 }
